+++ app/src/main/java/io/github/w1ll_du/MCUtilsBot/command/commands/ticksToHours.java
@@ -10,7 +10,7 @@ public class ticksToHours extends AConversionCommand {
     }
 
     @Override
-    public double convert(int v) {
-         return v / 72000.0;
+    public double convert(double v) {
+         return v / 72000;
     }
 }
